requirements.txt
Flask
pytesseract
Pillow
tensorflow
numpy
opencv-python
matplotlib
colorthief
gunicorn
